// src/app/api/model/deletion-blocker.ts
// @ts-nocheck
/**
 * CheckItOut Platform API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DeletionBlockerCategory } from './deletion-blocker-category';


export interface DeletionBlocker { 
    category?: DeletionBlockerCategory;
    count?: number;
    description?: string;
    entityDescription?: string;
    entityIds?: Array<number>;
    entityType?: string;
    reason?: string;
}